requirements.txt
aws-cdk-lib==2.30.0
constructs>=10.0.0,<11.0.0
pydantic>=1.10.2